функциональный компонент, начало верстки
