--- homework_15/src/components/add/Add.css
@@ -1,33 +1,0 @@
-.container__add
-{
-    background: #4a4e57;
-    border-style: solid;
-    border-radius: 1px;
-    border-color: #6e6e6e;
-    margin: 5px;
-    padding: 15px;
-    width: 500px;
-    box-shadow: 3px 3px 3px black;
-}
-
-.container__append
-{
-    display: flex;
-    flex-direction: column;
-}
-
-.container__append textarea
-{
-    padding: 10px 10px 0 10px;
-    margin: auto;
-    width: 95%;
-    height: 130px;
-}
-
-.container__append input
-{
-    margin: 10px 0 0 15px;
-    width: 95%;
-    height: 26px;
-    background: #ffffff;
-}
